Shutter-draj/requirements.txt
Django==1.8.4
django-braces==1.8.1
django-filter==0.11.0
django-oauth-toolkit==0.9.0
django-rest-framework-social-oauth2==1.0.2
djangorestframework==3.2.4
djoser==0.4.0
httpie==0.9.2
Markdown==2.6.2
MySQL-python==1.2.5
oauthlib==1.0.1
Pygments==2.0.2
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2016.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
django-storages==1.1.8
django-push-notifications==1.3.1
boto==2.40.0